unified project for MRA protocol
git-svn-id: http://svn.miranda-ng.org/main/trunk@13260 1316c22d-e87f-b044-9b9b-93d7a3e3ba9c

+++ protocols/MRA/src/MraProto.h
@@ -101,7 +101,8 @@ struct CMraProto : public PROTO<CMraProto>
 		return DB_SetStringExW(Contact, m_szModuleName, valueName, parValue); }
 
 	__forceinline bool mraWriteContactSettingBlob(MCONTACT hContact, const char *lpszValueName, LPVOID lpbValue, size_t dwValueSize) {
-		return db_set_blob(hContact, m_szModuleName, lpszValueName, lpbValue, dwValueSize) != 0; }
+		return db_set_blob(hContact, m_szModuleName, lpszValueName, lpbValue, (unsigned int)dwValueSize) != 0;
+	}
 	__forceinline bool mraGetContactSettingBlob(MCONTACT hContact, const char *lpszValueName, LPVOID lpbRet, size_t dwRetBuffSize, size_t *pdwRetBuffSize) {
 		return DB_GetContactSettingBlob(hContact, m_szModuleName, lpszValueName, lpbRet, dwRetBuffSize, pdwRetBuffSize); }
